Fix/budget/delete modal

+++ app/Http/Controllers/Budget/BudgetPeriodController.php
@@ -84,6 +84,8 @@ class BudgetPeriodController extends Controller
 
     public function store(Request $request)
     {
+        $this->authorize(PermissionEnum::BudgetSetupCreate, BudgetPeriods::class);
+
          $validated = $request->validate([
             'Name' => 'required|string|max:255',
             'FiscalYear' => 'required|integer|min:2020|max:2100',
@@ -153,6 +155,8 @@ class BudgetPeriodController extends Controller
 
     public function show($id)
     {
+        $this->authorize(PermissionEnum::BudgetSetupView, BudgetPeriods::class);
+
         try {
             // Fetch the budget
             $budget = Budget::findOrFail($id);
@@ -186,7 +190,7 @@ class BudgetPeriodController extends Controller
 
     public function update(Request $request, $id)
     {
-       // $this->authorize(PermissionEnum::BudgetSetupUpdate, BudgetPeriods::class);
+        $this->authorize(PermissionEnum::BudgetSetupUpdate, BudgetPeriods::class);
 
         $validated = $request->validate([
             'Name'        => 'required|string|max:255',
@@ -234,6 +238,7 @@ class BudgetPeriodController extends Controller
         $this->authorize(PermissionEnum::BudgetSetupDelete, BudgetPeriods::class);
         try{
             $period = BudgetPeriods::findOrFail($id); // safer: throws 404 if not found
+            $period -> DeletedBy = Auth::Id();
             $period->delete();
 
             activity()
@@ -252,14 +257,16 @@ class BudgetPeriodController extends Controller
 
     public function delGLAttachment($id)
     {
-        //$this->authorize(PermissionEnum::BudgetSetupDelete, BudgetGLsAttachments::class);
+        $this->authorize(PermissionEnum::BudgetSetupDelete, BudgetGLsAttachments::class);
 
         try {
             $glAttachment = BudgetGLsAttachments::findOrFail($id);
+            $glAttachment->DeletedBy = Auth::Id();
+            $glAttachment->save();
             $glAttachment->delete();
 
             //Delete the associated BudgetGLMasterAllocations if they exist
-            BudgetGLMasterAllocations::where('GLAttachmentID', $id)->delete();
+            // BudgetGLMasterAllocations::where('GLAttachmentID', $id)->delete();
 
             activity()
                 ->performedOn($glAttachment)
